+++ Entregas/02_matrices_dinamicas/matrixDin.h
@@ -7,7 +7,7 @@
 typedef struct matrixT
 {
     int dimFil, dimCol;
-    int** data;
+    float** data;
 } matrixT;
 
 
@@ -15,6 +15,6 @@ matrixT matrixLoader(FILE* fptr);
 void printMatrix(matrixT matrix);
 matrixT addMat(matrixT A, matrixT B);
 void freeMat(matrixT mat);
-void escalarMat(matrixT mat, int escalar);
+void escalarMat(matrixT mat, float escalar);
 
 #endif
